Improve modularity of code.
Previously, SPPT could process just Python applications. Although that
was enough for its original purpose (an experiment on TDD), it cannot
support new requirements, such as support to evaluate JFlap applications.

Thus, in order to support this, I reviewed the source code SPPT, moving
code (and rewrite some) to improve cohesion (coupling was already good
at that time).

The main improvement for this iteration is that SPPT supports multiple
graders, which are discovered at runtime. Given a submission, it can be
fed to each grader, which can inform whether it supports evaluating such
submission. If so, that grader will be executed. Currently, SPPT supports
several graders for the same submission (although we do not have multiple
graders available for now).

The 'Assessment' class is now more flexible, supporting nest of assignment
results.

The file 'index_new.php' was removed, as it was the original version of
SPPT, before converting it to an object oriented design.